refactor: refactore some code. change logo
