+++ frontend/src/app/app.html
@@ -1,4 +1,6 @@
 <tui-root [attr.tuiTheme]="themeService.theme()">
   <app-header></app-header>
-  <router-outlet />
+  <main>
+    <router-outlet />
+  </main>
 </tui-root>
