--- frontend/src/app/core/components/laguage-switcher/laguage-switcher.html
@@ -1,15 +1,0 @@
-<ng-container>
-  <button
-    appearance="primary"
-    iconStart="@tui.languages"
-    size="s"
-    tuiButton
-    tuiButtonSelect
-    type="button"
-    [(ngModel)]="value"
-    (ngModelChange)="onChangeLanguage($event)"
-  >
-    {{ value }}
-    <tui-data-list-wrapper *tuiTextfieldDropdown new [items]="languages" />
-  </button>
-</ng-container>
