+++ frontend/src/app/core/components/language-switcher/language-switcher.html
@@ -1,0 +1,13 @@
+<button
+  appearance="primary"
+  iconStart="@tui.languages"
+  size="s"
+  tuiButton
+  tuiButtonSelect
+  type="button"
+  [(ngModel)]="value"
+  (ngModelChange)="onChangeLanguage($event)"
+>
+  {{ value }}
+  <tui-data-list-wrapper *tuiTextfieldDropdown new [items]="languages" />
+</button>
