+++ frontend/src/app/core/components/language-switcher/language-switcher.scss
