+++ frontend/src/app/core/components/theme-switcher/theme-switcher.html
@@ -1,11 +1,9 @@
-<ng-container>
-  <input
-    size="s"
-    tuiLike
-    type="checkbox"
-    [checkedIcon]="'@tui.moon-filled'"
-    [uncheckedIcon]="'@tui.sun-filled'"
-    (change)="onChangeTheme()"
-    appearance="primary"
-  />
-</ng-container>
+<input
+  size="s"
+  tuiLike
+  type="checkbox"
+  [checkedIcon]="'@tui.moon-filled'"
+  [uncheckedIcon]="'@tui.sun-filled'"
+  (change)="onChangeTheme()"
+  appearance="primary"
+/>
