+++ frontend/src/app/core/layout/header/header.html
@@ -1,13 +1,11 @@
-<ng-container>
-  <header class="header" tuiHeader tuiAppearance="secondary-grayscale">
-    <div class="wrapper header__wrapper">
-      <a class="logo" routerLink="/">
-        <img class="logo_image" src="assets/images/logo.svg" alt="logo" />
-      </a>
-      <div class="services">
-        <app-theme-switcher></app-theme-switcher>
-        <app-laguage-switcher></app-laguage-switcher>
-      </div>
+<header class="header" tuiHeader tuiAppearance="secondary-grayscale">
+  <div class="wrapper header__wrapper">
+    <a class="logo" routerLink="/">
+      <img class="logo_image" src="assets/images/logo.svg" alt="logo" />
+    </a>
+    <div class="services">
+      <app-theme-switcher></app-theme-switcher>
+      <app-language-switcher></app-language-switcher>
     </div>
-  </header>
-</ng-container>
+  </div>
+</header>
