+++ frontend/src/app/core/layout/header/header.scss
@@ -14,11 +14,11 @@
 }
 
 .logo {
-  width: 100px;
+  width: 70px;
 }
 
 .logo_image {
-  width: 100px;
+  width: 70px;
 }
 
 .services {
