+++ frontend/src/app/shared/constants/storage-constants.ts
@@ -1,0 +1,3 @@
+export const STORAGE_KEYS = {
+  THEME: 'app-theme',
+} as const;
